Define IMPROMPTU_BIN and source prompt.
We source the prompt so that impromptu can have access to the shell's
prompt variables (they'll be hidden otherwise). We define
IMPROMPTU_BIN so that the bin directory can be accessed in impromptu
(as it's now sourced), and easily overridden for testing.

Likewise, for impromptu to work, your bashrc, zshrc, etc must now
contain `source impromptu` instead of anything setting the PS1.

See Impromptu/impromptu#47.